@@ -11,8 +11,8 @@
     "jetifier": "^1.6.5",
     "lodash": "^4.17.11",
     "native-base": "^2.12.1",
-    "react": "16.6.1",
-    "react-native": "0.57.5",
+    "react": "^16.8.6",
+    "react-native": "^0.59.10",
     "react-native-axios": "^0.17.1",
     "react-native-ffmpeg": "^0.3.6",
     "react-native-fs": "^2.13.3",
@@ -28,13 +28,13 @@
     "react-native-video-controls": "^2.2.3",
     "react-native-volume-control": "^1.0.1",
     "react-navigation": "^3.8.1",
-    "react-redux": "^7.2.0",
+    "react-redux": "^7.2.2",
     "redux": "^4.0.5"
   },
   "devDependencies": {
     "babel-jest": "25.1.0",
     "jest": "25.1.0",
-    "metro-react-native-babel-preset": "0.58.0",
+    "metro-react-native-babel-preset": "0.51.0",
     "react-test-renderer": "16.6.1"
   },
   "jest": {
