@@ -35,7 +35,6 @@ class Playlists extends Component {
         }
     }
 
-
     componentDidMount() {
         NetInfo.getConnectionInfo().then(data => {
             if(data.type == "unknown" || data.type == "cellular") {
@@ -57,6 +56,7 @@ class Playlists extends Component {
                         self.setState({songs:JSONified})
                     }
                 }
+                console.log('Arbeloa');
                 console.log(JSONified)
             });
             self.setState({isLoadingLocalStorage: false});
@@ -91,7 +91,7 @@ class Playlists extends Component {
         else {
           return false;
         }
-      }
+    }
 
     _onRefresh = () => {
       const self = this;
@@ -141,13 +141,13 @@ class Playlists extends Component {
     render() {
         const { navigation } = this.props;
         const item = navigation.getParam('item', 'ERROR');
-        const myIcon = (<Icon name={"ios-more"} size={25} color={"white"} style={{padding:12}} />)
+        const myIcon = (<Icon name={"ios-more"} size={25} color={"#444"} style={{padding:12}} />)
       return (
-        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'flex-start', backgroundColor: 'black' }}>
+        <View style={{ flex: 1, alignItems: 'center', justifyContent: 'flex-start', backgroundColor: '#D3E0EC' }}>
             {this.state.modalOpen ? (<ModalPlaylists deletePlaylist={this.deletePlaylist} playlist={this.state.currentPL} closeModal={this.closeModalEdit} />) : false}
             <View style={{width:'100%', height:'10%', flexDirection:'row', alignItems:'center', justifyContent:'flex-start'}}>
                 <View style={{width:'10%', height:'100%', alignItems:'center', justifyContent:'center'}}>
-                    <Icon name={"ios-search"} size={20} color={"white"} />
+                    <Icon name={"ios-search"} size={20} color={"#444"} />
                 </View>
                 <TextInput
                     onChangeText={(text) => {
@@ -156,14 +156,14 @@ class Playlists extends Component {
                     }
                     }
                     placeholder={"Search playlist"}
-                    placeholderTextColor={"white"}
+                    placeholderTextColor={"#444"}
                     value={this.state.searchPlaylistText}
-                    style={{width:'90%', height:'100%', color:'white'}} />
+                    style={{width:'90%', height:'100%', color:'#444'}} />
             </View>
             <View style={{width:'100%', height:'14%', flexDirection:'row', alignItems:'flex-end', justifyContent:'space-between', paddingLeft:'5%', paddingRight:'5%',}}>
                 <View>
-                    <Text style={{color:'white', marginBottom:10, fontSize:13, fontWeight:'bold'}}>PLAYLISTS</Text>
-                    <Text style={{color:'white', fontSize:20, marginBottom:10}}>Your playlists</Text>
+                    <Text style={{color:'#444', marginBottom:10, fontSize:13, fontWeight:'bold'}}>PLAYLISTS</Text>
+                    <Text style={{color:'#444', fontSize:20, marginBottom:10}}>Your playlists</Text>
                 </View>
                 <TouchableOpacity
                 style={{marginBottom:10}}
@@ -199,7 +199,7 @@ class Playlists extends Component {
             onRefresh={this._onRefresh}
           />
         }
-        style={{width:'100%', backgroundColor:'black'}}>
+        style={{width:'100%', backgroundColor:'#D3E0EC'}}>
             {this.state.displayPlaylists.map((play, index) => 
                 play == "songsDownloadedOnDevice" ?
                 null
@@ -207,16 +207,10 @@ class Playlists extends Component {
                 (
                 <TouchableOpacity onLongPress={() => this.openModalEdit(play)} onPress={() => this.props.navigation.navigate('PlaylistSongs', {key: play, comeFrom: 'Playlists'})} key={play} style={{backgroundColor:'#1A237E', width:'90%', marginLeft:'5%', height: 110, borderRadius: 13, overflow:'hidden', marginBottom:12, marginTop:12}}>
                     <ImageBackground style={{width:'100%', height:'100%'}} resizeMode={"cover"} source={{uri: play === "favorites__Playlist" ? this.state.images[index%this.state.images.length] : this.getPlaylistBackground(play) ? this.getPlaylistBackground(play) : this.state.images[index%this.state.images.length]}} >
-                        <View style={{width:'100%', height:'100%', backgroundColor: 'rgba(0, 0, 0, 0.6)', position: 'absolute'}}>
+                        <View style={{width:'100%', height:'100%', backgroundColor: 'rgba(234,76,137, 0.45)', position: 'absolute'}}>
                             <Text style={{color:'white', fontSize:24, position:'absolute', top:10, left:20, fontWeight:'bold'}}>{play === "favorites__Playlist" ? "FAVORITES" :play.toUpperCase()}</Text>
                         </View>
                     </ImageBackground>
-                    {/*<OptionsMenu
-                        customButton={myIcon}
-                        destructiveIndex={1}
-                        options={["Eliminar"]}
-                        actions={[() => this.deletePlaylist(play)]}
-                    />*/}
                 </TouchableOpacity>
                 )
             )}
@@ -239,12 +233,12 @@ class Playlists extends Component {
         </View>
       );
     }
+
     goBack = () => {
         this.props.navigation.goBack()
     }
 
     deletePlaylist = (key) => {
-        console.log(key);
         let jsonPlaylists = null;
         let jsonSongs = null;
         AsyncStorage.multiGet(["Playlists", "Songs"], (err, stores) => {
@@ -327,7 +321,6 @@ class Playlists extends Component {
                         if(!isSavedAlready) newSongsArray.push(songObject);
                         console.log(newSongsArray)
                         this.props.screenProps.setLocalePlaylist(newSongsArray);
-                        //songs.push(songObject)
                         self.props.screenProps.addToAllSongs(newSongsArray);
                         songs = JSON.stringify(newSongsArray);
                         AsyncStorage.setItem("Songs", songs)
@@ -398,7 +391,6 @@ class Playlists extends Component {
                 .catch(error => console.log(error));
             }
             else {
-                console.log("Ya existe")
             }
           }
           catch(error) {
