@@ -1,5 +1,9 @@
 import React, { Component } from 'react';
-import { ImageBackground, View, Text } from 'react-native';
+import { Image, View, Dimensions, StyleSheet } from 'react-native';
+const screenWidth = Dimensions.get('screen').width;
+
+const thumbnailSize = (screenWidth/100)*80;
+const thumbnailMargin = (100 - thumbnailSize) / 2;
 
 class RelatedThumbnails extends Component {
     constructor(props){
@@ -8,15 +12,23 @@ class RelatedThumbnails extends Component {
     }
     render() {
       const { styling, src, isAudioSource, item } = this.props;
+      const style = {
+        width: thumbnailSize,
+        height: thumbnailSize,
+        marginRight: thumbnailMargin,
+        marginLeft: thumbnailMargin,
+        borderRadius: 6,
+        justifyContent:'center',
+        alignItems:'center'
+      };
       return ( 
-        <ImageBackground style={styling} resizeMode={"cover"} source={src}>
-          <View style={{width: '100%', height: '100%', backgroundColor:'rgba(0, 0, 0, 0.6)', paddingBottom:'6%', alignItems:'center', justifyContent:'flex-end'}}>
-            <View style={{width:'90%', height: '20%', alignItems: 'center', justifyContent: 'space-between',}}>
-              <Text style={{width:'100%', fontSize:16, textAlign:'center', color:'white'}} numberOfLines={1} ellipsizeMode={"tail"}>{item === null ? this.props.videoTitle : !item.customName ? this.props.videoTitle : item.customName}</Text>
-              <Text style={{width:'100%', fontSize:13, textAlign:'center', color:'white', fontWeight:'bold'}} numberOfLines={1} ellipsizeMode={"tail"}>{item === null ? this.props.videoChannel : !item.customArtist ? this.props.videoChannel : item.customArtist}</Text>
-            </View>
-          </View>
-        </ImageBackground>
+      <View style={style}>
+        <Image
+          style={{position:'absolute', top:0, left:0, width:'100%', height:'100%', borderRadius: 6}}
+          resizeMode={"cover"}
+          source={src}
+        />
+      </View>
       );
     }
 }
