@@ -1,7 +1,18 @@
 /** @format */
 
+import React from 'react';
 import {AppRegistry} from 'react-native';
+import { Provider } from 'react-redux';
 import App from './App';
 import {name as appName} from './app.json';
+import store from './redux/store/store';
 
-AppRegistry.registerComponent(appName, () => App);
+const AppComponent = () => {
+    return(
+        <Provider store={store}>
+            <App />
+        </Provider>
+    );
+}
+
+AppRegistry.registerComponent(appName, () => AppComponent);
